+++ src/services/CertificatePayloadGenerator.ts
@@ -2,7 +2,7 @@ import { Inject, Service } from 'typedi';
 import { CERTIFICATE_DATA } from '../models/Enums';
 import { ITestResult } from '../models/ITestResult';
 import { CertificatePayloadGeneratorIva } from './certificate-payload/CertificatePayloadGeneratorIva';
-import { CertificatePayloadGeneratorAdr } from './CertificatePayloadGeneratorAdr';
+import { CertificatePayloadGeneratorAdr } from './certificate-payload/CertificatePayloadGeneratorAdr';
 import { CertificatePayloadGeneratorRwt } from './certificate-payload/CertificatePayloadGeneratorRwt';
 import { CertificatePayloadGeneratorPassOrFail } from './CertificatePayloadGeneratorPassOrFail';
 import { CertificatePayloadGeneratorMsva } from './certificate-payload/CertificatePayloadGeneratorMsva';
